+++ test/test_publish.py
@@ -178,6 +178,19 @@ def _graph_path(dvcs_type):
         master = "default"
     return join("graphs", "branch-{0}".format(master), "machine-tarzan", "time_func.json")
 
+
+def test_publish_range_spec(generate_result_dir):
+    conf, repo, commits = generate_result_dir(5 * [1])
+    for range_spec, expected in (
+        ([commits[0], commits[-1]], set([commits[0], commits[-1]])),
+        ('HEAD~2..HEAD' if repo.dvcs == 'git' else '.~1:',
+            set(commits[-2:])),
+    ):
+        tools.run_asv_with_conf(conf, "publish", range_spec)
+        data = util.load_json(join(conf.html_dir, 'index.json'))
+        assert set(data['revision_to_hash'].values()) == expected
+
+
 def test_regression_simple(generate_result_dir):
     conf, repo, commits = generate_result_dir(5 * [1] + 5 * [10])
     tools.run_asv_with_conf(conf, "publish")