+++ test/test_repo.py
@@ -155,7 +155,7 @@ def test_repo_hg(tmpdir):
     conf = config.Config()
 
     dvcs = tools.generate_test_repo(tmpdir, list(range(10)), dvcs_type='hg', 
-                                    extra_branches=[('default~4', 'some-branch',[11, 12, 13])])
+                                    extra_branches=[('default~4', 'somebranch',[11, 12, 13])])
 
     mirror_dir = join(tmpdir, "repo")
 
@@ -166,10 +166,10 @@ def test_repo_hg(tmpdir):
                            master="default", branch="tag5",
                            is_remote=is_remote)
 
-        conf.branches = ['default', 'some-branch']
+        conf.branches = ['default', 'somebranch']
         branch_commits = {
             'default': [dvcs.get_hash('default'), dvcs.get_hash('default~6')],
-            'some-branch': [dvcs.get_hash('some-branch'), dvcs.get_hash('some-branch~6')]
+            'somebranch': [dvcs.get_hash('somebranch'), dvcs.get_hash('somebranch~6')]
         }
         _test_branches(conf, branch_commits)
 
@@ -397,3 +397,40 @@ def test_git_submodule(tmpdir):
     r.checkout(checkout_dir, commit_hash_5)
     assert os.path.isfile(join(checkout_dir, 'README'))
     assert not os.path.isdir(join(checkout_dir, 'sub1'))
+
+
+@pytest.mark.parametrize('dvcs_type', [
+    "git",
+    pytest.mark.skipif(hglib is None, reason="needs hglib")("hg")
+])
+def test_root_ceiling(dvcs_type, tmpdir):
+    # Check that git/hg does not try to look for repository in parent
+    # directories.
+    tmpdir = six.text_type(tmpdir)
+    dvcs1 = tools.generate_repo_from_ops(tmpdir, dvcs_type, [("commit", 1)])
+    dvcs2 = tools.generate_repo_from_ops(tmpdir, dvcs_type, [("commit", 2)])
+    commit1 = dvcs1.get_branch_hashes()[0]
+    commit2 = dvcs2.get_branch_hashes()[0]
+
+    conf = config.Config()
+    conf.branches = []
+    conf.dvcs = dvcs_type
+    conf.project = join(tmpdir, "repo")
+    conf.repo = dvcs1.path
+
+    r = repo.get_repo(conf)
+
+    # Checkout into a subdir inside another repository
+    workcopy_dir = join(dvcs2.path, "workcopy")
+    r.checkout(workcopy_dir, commit1)
+
+    # Corrupt the checkout
+    for pth in ['.hg', '.git']:
+        pth = os.path.join(workcopy_dir, pth)
+        if os.path.isdir(pth):
+            shutil.rmtree(pth)
+
+    # Operation must fail (commit2 is not in dvcs1), not use the
+    # parent repository
+    with pytest.raises(Exception):
+        r.checkout(workcopy_dir, commit2)